services/registration_service/registration_service/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0